[Fix] 테스트파일 삭제
